// apps/worker/ui/src/api/types.ts
export type GetListRequest = {
  offset: number;
  limit: number | null;
  selected_ids: number[];
  search_text: string | null;
  text_filter: Record<string, string[]>;
  filter: Record<string, unknown[]>;
  sort: [string, 'asc' | 'desc'] | null;
  random?: boolean;
};
// ...
export type ImageRecord = {
  id?: number;
  created_at?: string | null;
  updated_at?: string | null;
  prompt?: string | null;
  negative_prompt?: string | null;
  prompt_embedding?: number[] | null;
  object_key?: string | null;
};